backend/requirements.txt
fastapi
uvicorn
python-multipart
python-pptx
google-generativeai
edge-tts
aiofiles
python-dotenv
